fix type for cu_seqlen, uninstall causal_conv1d

+++ src/axolotl/monkeypatch/models/qwen3_5/modeling.py
@@ -35,7 +35,7 @@ def get_cu_seqlens(position_ids):
     if position_ids.ndim == 3:
         position_ids = position_ids[0]
 
-    tensor_kwargs = {"dtype": torch.long, "device": position_ids.device}
+    tensor_kwargs = {"dtype": torch.int32, "device": position_ids.device}
     position_ids = position_ids.reshape(-1)
     indices_q = (position_ids == 0).nonzero().reshape(-1)
     return torch.cat(
